Moving block helpers to utils directory
