+++ src/app/components/Image/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { imagePropTypes, imageDefaultPropTypes } from '../../proptypes';
-import filterForBlockType from '../../BlockHelpers/blockHelpers';
+import filterForBlockType from '../../../utils/blockHelpers';
 import { FF_NEWS_SANS_REG } from '../../../lib/constants/styles';
 
 const getText = ({ model }) => model.blocks[0].model.blocks[0].model.text;
